requirements.txt
Django~=3.1.0
djangorestframework==3.12.1
djangofilters==2.4.0
